+++ packages/core/@generated/graphql/index.ts
@@ -18,51 +18,78 @@ export type Scalars = {
   Float: number
 }
 
+/** Shopping cart identification input. */
 export type IStoreCart = {
+  /** Order information, including `orderNumber` and `acceptedOffer`. */
   order: IStoreOrder
 }
 
+/** Image input. */
 export type IStoreImage = {
+  /** Alias for the input image. */
   alternateName: Scalars['String']
+  /** Image input URL. */
   url: Scalars['String']
 }
 
+/** Offer input. */
 export type IStoreOffer = {
+  /** Information on the item being offered. */
   itemOffered: IStoreProduct
+  /** This is displayed as the "from" price in the context of promotions' price comparison. This may change before it reaches the shelf. */
   listPrice: Scalars['Float']
+  /** Also known as spot price. */
   price: Scalars['Float']
+  /** Number of items offered. */
   quantity: Scalars['Int']
+  /** Seller responsible for the offer. */
   seller: IStoreOrganization
 }
 
+/** Offer input. */
 export type IStoreOrder = {
+  /** Array with information on each accepted offer. */
   acceptedOffer: Array<IStoreOffer>
+  /** Order shopping cart ID, also known as `orderFormId`. */
   orderNumber: Scalars['String']
 }
 
+/** Organization input. */
 export type IStoreOrganization = {
+  /** Organization ID. */
   identifier: Scalars['String']
 }
 
+/** Product input. */
 export type IStoreProduct = {
+  /** Array of product images. */
   image: Array<IStoreImage>
+  /** Product name. */
   name: Scalars['String']
+  /** Stock Keeping Unit ID. */
   sku: Scalars['String']
 }
 
+/** Selected facet input. */
 export type IStoreSelectedFacet = {
   key: Scalars['String']
   value: Scalars['String']
 }
 
+/** Session input. */
 export type IStoreSession = {
+  /** Session input channel. */
   channel: InputMaybe<Scalars['String']>
+  /** Session input country. */
   country: InputMaybe<Scalars['String']>
+  /** Session input postal code. */
   postalCode: InputMaybe<Scalars['String']>
 }
 
 export type Mutation = {
+  /** Update session information. */
   updateSession: StoreSession
+  /** Returns the order if anything has changed in it, or `null` if the order is valid. */
   validateCart: Maybe<StoreCart>
 }
 
@@ -109,177 +136,293 @@ export type QuerySearchArgs = {
   term?: InputMaybe<Scalars['String']>
 }
 
+/** Aggregate offer information, for a given SKU that is available to be fulfilled by multiple sellers. */
 export type StoreAggregateOffer = {
+  /** Highest price among all sellers. */
   highPrice: Scalars['Float']
+  /** Lowest price among all sellers. */
   lowPrice: Scalars['Float']
+  /** Number of sellers selling this SKU. */
   offerCount: Scalars['Int']
+  /** Array with information on each available offer. */
   offers: Array<StoreOffer>
+  /** ISO code of the currency used for the offer prices. */
   priceCurrency: Scalars['String']
 }
 
+/** Average rating, based on multiple ratings or reviews. */
 export type StoreAggregateRating = {
+  /** Value of the aggregate rating. */
   ratingValue: Scalars['Float']
+  /** Total number of ratings. */
   reviewCount: Scalars['Int']
 }
 
+/** information about the author of a product review or rating. */
 export type StoreAuthor = {
+  /** Author name. */
   name: Scalars['String']
 }
 
+/** Brand of a given product. */
 export type StoreBrand = {
+  /** Brand name. */
   name: Scalars['String']
 }
 
+/** List of items consisting of chain linked web pages, ending with the current page. */
 export type StoreBreadcrumbList = {
+  /** Array with breadcrumb elements. */
   itemListElement: Array<StoreListItem>
+  /** Number of breadcrumbs in the list. */
   numberOfItems: Scalars['Int']
 }
 
+/** Shopping cart information. */
 export type StoreCart = {
+  /** List of shopping cart messages. */
   messages: Array<StoreCartMessage>
+  /** Order information, including `orderNumber` and `acceptedOffer`. */
   order: StoreOrder
 }
 
+/** Shopping cart message. */
 export type StoreCartMessage = {
+  /** Shopping cart message status, which can be `INFO`, `WARNING` OR `ERROR`. */
   status: StoreStatus
+  /** Shopping cart message text. */
   text: Scalars['String']
 }
 
+/** Product collection information. */
 export type StoreCollection = {
+  /** List of items consisting of chain linked web pages, ending with the current page. */
   breadcrumbList: StoreBreadcrumbList
+  /** Collection ID. */
   id: Scalars['ID']
+  /** Collection meta information. Used for search. */
   meta: StoreCollectionMeta
+  /** Meta tag data. */
   seo: StoreSeo
+  /** Corresponding collection URL slug, with which to retrieve this entity. */
   slug: Scalars['String']
+  /** Collection type. */
   type: StoreCollectionType
 }
 
+/** Collection connection pagination information. */
 export type StoreCollectionConnection = {
+  /** Array with collection connection page edges. */
   edges: Array<StoreCollectionEdge>
+  /** Collection connection page information. */
   pageInfo: StorePageInfo
 }
 
+/** Collection pagination edge. */
 export type StoreCollectionEdge = {
+  /** Collection pagination cursor. */
   cursor: Scalars['String']
+  /** Collection pagination node. */
   node: StoreCollection
 }
 
+/** Product collection facet, used for search. */
 export type StoreCollectionFacet = {
+  /** Facet key. */
   key: Scalars['String']
+  /** Facet value. */
   value: Scalars['String']
 }
 
+/** Collection meta information. Used for search. */
 export type StoreCollectionMeta = {
+  /** List of selected collection facets. */
   selectedFacets: Array<StoreCollectionFacet>
 }
 
+/** Product collection type. Possible values are `Department`, `Category`, `Brand` or `Cluster`. */
 export type StoreCollectionType =
   | 'Brand'
   | 'Category'
   | 'Cluster'
   | 'Department'
 
+/** Search facet information. */
 export type StoreFacet = {
+  /** Facet key. */
   key: Scalars['String']
+  /** Facet label. */
   label: Scalars['String']
+  /** Facet type. Possible values are `BOOLEAN` and `RANGE`. */
   type: StoreFacetType
+  /** Array with information on each facet value. */
   values: Array<StoreFacetValue>
 }
 
+/** Search facet type. */
 export type StoreFacetType = 'BOOLEAN' | 'RANGE'
 
+/** Information of a specific facet value. */
 export type StoreFacetValue = {
+  /** Facet value label. */
   label: Scalars['String']
+  /** Number of items with this facet. */
   quantity: Scalars['Int']
+  /** Indicates whether facet is selected. */
   selected: Scalars['Boolean']
+  /** Facet value. */
   value: Scalars['String']
 }
 
+/** Image. */
 export type StoreImage = {
+  /** Alias for the image. */
   alternateName: Scalars['String']
+  /** Image URL. */
   url: Scalars['String']
 }
 
+/** Item of a list. */
 export type StoreListItem = {
+  /** List item value. */
   item: Scalars['String']
+  /** Name of the list item. */
   name: Scalars['String']
+  /** Position of the item in the list. */
   position: Scalars['Int']
 }
 
+/** Offer information. */
 export type StoreOffer = {
+  /** Offer item availability. */
   availability: Scalars['String']
+  /** Offer item condition. */
   itemCondition: Scalars['String']
+  /** Information on the item being offered. */
   itemOffered: StoreProduct
+  /** This is displayed as the "from" price in the context of promotions' price comparison. This may change before it reaches the shelf. */
   listPrice: Scalars['Float']
+  /** Also known as spot price. */
   price: Scalars['Float']
+  /** ISO code of the currency used for the offer prices. */
   priceCurrency: Scalars['String']
+  /** Next date in which price is scheduled to change. If there is no scheduled change, this will be set a year in the future from current time. */
   priceValidUntil: Scalars['String']
+  /** Number of items offered. */
   quantity: Scalars['Int']
+  /** Seller responsible for the offer. */
   seller: StoreOrganization
+  /** Computed price before applying coupons, taxes or benefits. This may change before it reaches the shelf. */
   sellingPrice: Scalars['Float']
 }
 
+/** Information of a specific order. */
 export type StoreOrder = {
+  /** Array with information on each accepted offer. */
   acceptedOffer: Array<StoreOffer>
+  /** Order shopping cart ID, also known as `orderFormId`. */
   orderNumber: Scalars['String']
 }
 
+/** Organization. */
 export type StoreOrganization = {
+  /** Organization ID. */
   identifier: Scalars['String']
 }
 
+/** Page information. */
 export type StorePageInfo = {
+  /** Page cursor end. */
   endCursor: Scalars['String']
+  /** Indicates whether next page exists. */
   hasNextPage: Scalars['Boolean']
+  /** Indicates whether previous page exists. */
   hasPreviousPage: Scalars['Boolean']
+  /** Page cursor start. */
   startCursor: Scalars['String']
+  /** Total number of items (products or collections), not pages. */
   totalCount: Scalars['Int']
 }
 
+/** Client profile data. */
 export type StorePerson = {
+  /** Client email. */
   email: Scalars['String']
+  /** Client last name. */
   familyName: Scalars['String']
+  /** Client first name. */
   givenName: Scalars['String']
+  /** Client ID. */
   id: Scalars['String']
 }
 
+/** Product information. */
 export type StoreProduct = {
+  /** Array of additional properties. */
   additionalProperty: Array<StorePropertyValue>
+  /** Aggregate ratings data. */
   aggregateRating: StoreAggregateRating
+  /** Product brand. */
   brand: StoreBrand
+  /** List of items consisting of chain linked web pages, ending with the current page. */
   breadcrumbList: StoreBreadcrumbList
+  /** Product description. */
   description: Scalars['String']
+  /** Global Trade Item Number. */
   gtin: Scalars['String']
+  /** Array of images. */
   image: Array<StoreImage>
+  /** Indicates product group related to this product. */
   isVariantOf: StoreProductGroup
+  /** Product name. */
   name: Scalars['String']
+  /** Aggregate offer information. */
   offers: StoreAggregateOffer
+  /** Product ID. */
   productID: Scalars['String']
+  /** Array with review information. */
   review: Array<StoreReview>
+  /** Meta tag data. */
   seo: StoreSeo
+  /** Stock Keeping Unit ID. */
   sku: Scalars['String']
+  /** Corresponding collection URL slug, with which to retrieve this entity. */
   slug: Scalars['String']
 }
 
+/** Product connection pagination information. */
 export type StoreProductConnection = {
+  /** Array with product connection page edges. */
   edges: Array<StoreProductEdge>
+  /** Product connection page information. */
   pageInfo: StorePageInfo
 }
 
+/** Product pagination edge. */
 export type StoreProductEdge = {
+  /** Product pagination cursor. */
   cursor: Scalars['String']
+  /** Product pagination node. */
   node: StoreProduct
 }
 
+/** Product group information. */
 export type StoreProductGroup = {
+  /** Array of additional properties. */
   additionalProperty: Array<StorePropertyValue>
+  /** Array of variants related to product group. */
   hasVariant: Array<StoreProduct>
+  /** Product group name. */
   name: Scalars['String']
+  /** Product group ID. */
   productGroupID: Scalars['String']
 }
 
+/** Properties that can be associated with products and products groups. */
 export type StorePropertyValue = {
+  /** Property name. */
   name: Scalars['String']
+  /** Property value. */
   value: Scalars['String']
 }
 
@@ -293,9 +436,14 @@ export type StoreReviewRating = {
   ratingValue: Scalars['Float']
 }
 
+/** Search result. */
 export type StoreSearchResult = {
+  /** Array of search result facets. */
   facets: Array<StoreFacet>
+  /** Search result products. */
   products: StoreProductConnection
+  /** Search result suggestions. */
+  suggestions: StoreSuggestions
 }
 
 export type StoreSeo = {
@@ -305,12 +453,17 @@ export type StoreSeo = {
   titleTemplate: Scalars['String']
 }
 
+/** Session information. */
 export type StoreSession = {
+  /** Session channel. */
   channel: Maybe<Scalars['String']>
+  /** Session country. */
   country: Maybe<Scalars['String']>
+  /** Session postal code. */
   postalCode: Maybe<Scalars['String']>
 }
 
+/** Product sorting options used in search. */
 export type StoreSort =
   | 'discount_desc'
   | 'name_asc'
@@ -323,12 +476,12 @@ export type StoreSort =
 
 export type StoreStatus = 'ERROR' | 'INFO' | 'WARNING'
 
-export type UpdateSessionMutationMutationVariables = Exact<{
-  session: IStoreSession
-}>
-
-export type UpdateSessionMutationMutation = {
-  updateSession: { channel: string | null }
+/** Suggestions information. */
+export type StoreSuggestions = {
+  /** Array with suggestion products' information. */
+  products: Maybe<Array<StoreProduct>>
+  /** Array with suggestion terms. */
+  terms: Maybe<Array<Scalars['String']>>
 }
 
 export type ProductSummary_ProductFragment = {
@@ -352,6 +505,14 @@ export type ProductSummary_ProductFragment = {
   }
 }
 
+export type UpdateSessionMutationMutationVariables = Exact<{
+  session: IStoreSession
+}>
+
+export type UpdateSessionMutationMutation = {
+  updateSession: { channel: string | null }
+}
+
 export type Filter_FacetsFragment = {
   key: string
   label: string
@@ -362,6 +523,38 @@ export type Filter_FacetsFragment = {
     selected: boolean
     quantity: number
   }>
+}
+
+export type SearchSuggestionsQueryQueryVariables = Exact<{
+  term: Scalars['String']
+}>
+
+export type SearchSuggestionsQueryQuery = {
+  search: {
+    suggestions: {
+      terms: Array<string> | null
+      products: Array<{
+        slug: string
+        sku: string
+        name: string
+        gtin: string
+        id: string
+        brand: { name: string; brandName: string }
+        isVariantOf: { productGroupID: string; name: string }
+        image: Array<{ url: string; alternateName: string }>
+        offers: {
+          lowPrice: number
+          offers: Array<{
+            availability: string
+            price: number
+            listPrice: number
+            quantity: number
+            seller: { identifier: string }
+          }>
+        }
+      }> | null
+    }
+  }
 }
 
 export type ProductDetailsFragment_ProductFragment = {