+++ packages/core/src/components/sections/ProductGallery/ProductGalleryPage.tsx
@@ -41,7 +41,7 @@ function GalleryPage({ page, title, showSponsoredProducts = true }: Props) {
             page={page}
             pageSize={middleItemIndex}
           />
-          <div className="product-listing__results-sponsored">
+          <div data-fs-product-listing-sponsored>
             <h3>Sponsored</h3>
             {/*
               TODO: Refactor this bit of code